[MLIR][Torch] Fix padding for GlobalAvgPool conversion to AtenAvgPool2dOp
The semantics of the AveragePool operation is that for every spatial
dimension of the input tensor, there are 2 padding values provided
in the 'pad' attribute if provided at all (default to 0). The
conversion of GlobalAvgPool to AtenAvgPool2dOp only generated a
padding array with a default number of 0s equivalent to the number
of spatial dims instead of 2 * num_spatial_dims. This PR fixes
this mistake.

Additionally, we fix the assignment of the 'lowPaddingIncludingNC'
and 'highPaddingIncludingNC' variables in the 'padInputTensor'
function based on the proper length and ordering of the pads attribute.